+++ fre/src/main.rs
@@ -1,5 +1,6 @@
 use parse_js::ast::{Node, Syntax};
-use parse_js::cst::*;
+use parse_js::pure::{PureTree};
+// use parse_js::cst::*;
 use parse_js::loc::Loc;
 use parse_js::parse;
 
@@ -23,22 +24,17 @@ fn main() {
 
   let tree = parse(src.as_bytes()).unwrap();
 
-  let pure: PureTree = PureTree::new("toplevel", tree).into();
+  let pure: PureTree = PureTree::new(tree);
 
-  let pure2 = PureTree::new(
-    "aaa",
-    Node::new(
+  let node2 = Node::new(
       Loc(0, 0),
       Syntax::IdentifierExpr {
         name: "bbb".to_string(),
       },
-    ),
+    
   );
-  let syntax: SyntaxTree = pure.into();
 
-  
+  pure.replace_pure(node2);
 
-  syntax.replace_pure(pure2.into());
-
-  println!("{:#?}", syntax);
+  println!("{:#?}", pure);
 }
